+++ app/services/ticket_services.rb
@@ -1,8 +1,6 @@
 require "net/http"
 require "uri"
-require "csv"
-require "prawn"
-require "prawn/table"
+
 
 class TicketServices
   def self.create_report(ticket_params)
@@ -40,7 +38,7 @@ class TicketServices
         created_at: new_report.created_at
             }
       when "pdf"
-        pdf_data = generate_pdf(new_ticket_report, new_report)
+        pdf_data = GenerateFilesServices.generate_pdf(new_ticket_report, new_report)
 
         {
           ticket_report: {
@@ -59,52 +57,14 @@ class TicketServices
         success: true,
         ticket_report: new_ticket_report,
         report: new_report,
-        csv_data: generate_csv(new_ticket_report, new_report) # Generar CSV
+        csv_data: GenerateFilesServices.generate_csv(new_ticket_report, new_report) # Generar CSV
       }
       end
     end
   end
 
   private
-  def self.generate_csv(ticket_report, report)
-    CSV.generate(headers: true) do |csv|
-      csv << [ "Ticket Report ID", "Total Tickets", "Event ID", "Format", "Sold Tickets", "Date", "Created At" ]
 
-      csv << [
-        ticket_report.id,
-        ticket_report.capacity,
-        report.event_id,
-        report.format,
-        report.sold_tickets,
-        report.date,
-        report.created_at
-      ]
-    end
-  end
-
-  def self.generate_pdf(new_ticket_report, new_report)
-    Prawn::Document.new do |pdf|
-      pdf.text "Ticket Report", size: 20, style: :bold
-      pdf.move_down 20
-
-      # Generar la tabla
-      data = [
-        [ "Field", "Value" ],
-        [ "Event ID", new_report.event_id.to_s ],
-        [ "Total Tickets", new_ticket_report.capacity.to_s ],
-        [ "Sold Tickets", new_report.sold_tickets.to_s ],
-        [ "Date", new_report.date.to_s ],
-        [ "Created At", new_report.created_at.to_s ]
-      ]
-
-      pdf.table(data, header: true, row_colors: [ "dddddd", "ffffff" ], position: :center) do
-        cells.padding = 12
-        cells.borders = [ :bottom ]
-        cells.border_width = 1
-        row(0).font_style = :bold
-      end
-    end.render
-  end
 
   def get_ticket_summary
     uri = URI("asdas")
